requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bootstrap4==0.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
dj-static==0.0.6
Django==4.0.5
django-bootstrap==0.2.4
django-bootstrap-datepicker-plus==4.0.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
Pillow==9.1.1
python-dateutil==1.5
python-decouple==3.6
requests==2.28.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
urllib3==1.26.9
